+++ packages/react/src/components/Button/index.ts
@@ -1,9 +1,7 @@
 export { default as Button } from './Button';
-export { default as ButtonSkeleton } from './ButtonSkeleton';
 export { default as IconButton } from './IconButton';
 export { default as SearchButton } from './SearchButton';
 
 export * from './Button';
-export * from './ButtonSkeleton';
 export * from './IconButton';
 export * from './SearchButton';
