+++ packages/react/src/components/Skeleton/index.ts
@@ -1,0 +1,5 @@
+export { default as ButtonSkeleton } from './ButtonSkeleton';
+export { default as TextSkeleton } from './TextSkeleton';
+
+export * from './ButtonSkeleton';
+export * from './TextSkeleton';
